update for new Studio

+++ bin/run_tab.py
@@ -9,10 +9,12 @@ Dr. Paul Macklin (macklinp@iu.edu)
 import sys
 import os
 import time
+import logging
 from pathlib import Path
 from PyQt5 import QtCore, QtGui
 # from PyQt5.QtWidgets import *
 from PyQt5.QtWidgets import QFrame,QApplication,QWidget,QTabWidget,QFormLayout,QLineEdit, QHBoxLayout,QVBoxLayout,QRadioButton,QLabel,QCheckBox,QComboBox,QScrollArea, QPushButton,QPlainTextEdit
+from PyQt5.QtWidgets import QMessageBox
 
 from PyQt5.QtCore import QProcess
 
@@ -23,30 +25,36 @@ class QHLine(QFrame):
         self.setFrameShadow(QFrame.Sunken)
 
 class RunModel(QWidget):
-    def __init__(self, nanohub_flag, tab_widget, download_menu):
+    def __init__(self, nanohub_flag, tab_widget, rules_flag, download_menu):
         super().__init__()
 
         self.nanohub_flag = nanohub_flag
         self.tab_widget = tab_widget
+        self.rules_flag = rules_flag
         self.download_menu = download_menu
-        # self.legend_tab = legend_tab
 
         #-------------------------------------------
         # used with nanoHUB app
         # self.nanohub = True
+        # following set in pmb.py
+        self.current_dir = ''   
+        self.config_file = None
         self.tree = None
-        # following set in studio.py
-        self.homedir = ''   
+
+        # these get set in pmb.py
         self.config_tab = None
         self.microenv_tab = None
         self.celldef_tab = None
         self.user_params_tab = None
+        self.rules_tab = None
+        self.vis_tab = None
         self.legend_tab = None
 
-        #-----
-        self.vis_tab = None
+        self.output_dir = 'output'
 
+        #-----
         self.sim_output = QWidget()
+
         self.main_layout = QVBoxLayout()
 
         self.scroll = QScrollArea()
@@ -70,8 +78,8 @@ class RunModel(QWidget):
         self.run_button.clicked.connect(self.run_model_cb)
 
         self.cancel_button = QPushButton("Cancel")
-        # self.cancel_button.setStyleSheet("background-color: red")
-        self.cancel_button.setStyleSheet("background-color: rgb(250,100,100)")
+        self.cancel_button.setStyleSheet("background-color: red")
+        # self.cancel_button.setStyleSheet("background-color: rgb(250,50,50)")
         hbox.addWidget(self.cancel_button)
         self.cancel_button.clicked.connect(self.cancel_model_cb)
 
@@ -83,9 +91,10 @@ class RunModel(QWidget):
         self.exec_name = QLineEdit()
         if self.nanohub_flag:
             self.exec_name.setText('template')
+            self.exec_name.setEnabled(False)
         else:
-            self.exec_name.setText('../template')
-        # self.exec_name.setEnabled(False)
+            # self.exec_name.setText('../myproj')
+            self.exec_name.setText('template')
         hbox.addWidget(self.exec_name)
 
         hbox.addWidget(QLabel("Config:"))
@@ -119,60 +128,103 @@ class RunModel(QWidget):
 #------------------------------
     def update_xml_from_gui(self):
         self.xml_root = self.tree.getroot()
+        logging.debug(f'\n ========================== run_tab.py: update_xml_from_gui(): self.xml_root = {self.xml_root}')
         self.config_tab.xml_root = self.xml_root
         self.microenv_tab.xml_root = self.xml_root
         self.celldef_tab.xml_root = self.xml_root
         self.user_params_tab.xml_root = self.xml_root
+        if self.rules_flag:
+            self.rules_tab.xml_root = self.xml_root
+        # sys.exit(1)
 
         self.config_tab.fill_xml()
         self.microenv_tab.fill_xml()
         self.celldef_tab.fill_xml()
         self.user_params_tab.fill_xml()
+        if self.rules_flag:
+            self.rules_tab.fill_xml()
 
+        # self.vis_tab.circle_radius = ET.parse(self.xml_root)
+        # tree = ET.parse(xml_file)
+        # root = tree.getroot()
+        # rwh - warning: assumes "R_circle" name won't change
+        # self.vis_tab.mech_voxel_size = float(self.xml_root.find(".//user_parameters//mechanics_voxel_size").text)
+        # print("\n\n------------- run_tab(): self.vis_tab.mech_voxel_size = ",self.vis_tab.mech_voxel_size)
         
     def message(self, s):
         self.text.appendPlainText(s)
 
     def run_model_cb(self):
-        print("===========  run_model_cb():  ============")
+        logging.debug(f'===========  run_model_cb():  ============')
+
+        if float(self.config_tab.svg_interval.text()) != float(self.config_tab.full_interval.text()):
+            msg = QMessageBox()
+            msg.setIcon(QMessageBox.Critical)
+            msg.setText("Warning")
+            msg.setInformativeText('The output intervals for SVG and full (in Config Basics) do not match.')
+            msg.setWindowTitle("Warning")
+            msg.exec_()
+
         self.run_button.setEnabled(False)
 
         # if self.nanohub_flag: # copy normal workflow of an app, strange as it is
         if True: # copy normal workflow of an app, strange as it is
 
             # make sure we are where we started (app's root dir)
-            os.chdir(self.homedir)
+            logging.debug(f'\n------>>>> doing os.chdir to {self.current_dir}')
+            os.chdir(self.current_dir)
 
             # remove any previous data
             # NOTE: this dir name needs to match the <folder>  in /data/<config_file.xml>
-            os.system('rm -rf tmpdir*')
+            if self.nanohub_flag:
+                os.system('rm -rf tmpdir*')
+            # else:
+                # os.system('rm -rf output*')
             time.sleep(1)
-            if os.path.isdir('tmpdir'):
+            if self.nanohub_flag and s.path.isdir('tmpdir'):
                 # something on NFS causing issues...
                 tname = tempfile.mkdtemp(suffix='.bak', prefix='tmpdir_', dir='.')
                 shutil.move('tmpdir', tname)
-            os.makedirs('tmpdir')
+            if self.nanohub_flag:
+                os.makedirs('tmpdir')
+            else:
+                # os.makedirs('output')
+                self.output_dir = self.config_tab.folder.text()
+                # os.system('rm -rf tmpdir*')
+                os.system('rm -rf ' + self.output_dir)
+                logging.debug(f'run_tab.py:  doing: mkdir {self.output_dir}')
+                os.makedirs(self.output_dir)  # do 'mkdir output_dir'
+                time.sleep(1)
 
             # write the default config file to tmpdir
             # new_config_file = "tmpdir/config.xml"  # use Path; work on Windows?
-            tdir = os.path.abspath('tmpdir')
+            if self.nanohub_flag:
+                tdir = os.path.abspath('tmpdir')
+            else:
+                # tdir = os.path.abspath('.')
+                tdir = os.path.abspath(self.output_dir)
+
             new_config_file = Path(tdir,"config.xml")
+            self.celldef_tab.config_path = new_config_file
+            self.update_xml_from_gui()
 
             # write_config_file(new_config_file)  
             # update the .xml config file
-            self.update_xml_from_gui()
-
             # self.config_tab.fill_xml()
             # self.microenv_tab.fill_xml()
             # self.celldef_tab.fill_xml()
             # self.user_params_tab.fill_xml()
-            print("\n\n ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
-            print("run_tab.py: ----> writing modified model to ",new_config_file)
-            self.tree.write(new_config_file)  # saves modified XML to tmpdir/config.xml 
+            # print("\n\n ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
+            logging.debug(f'run_tab.py: ----> writing modified model to {self.config_file}')
+            # print("run_tab.py: ----> writing modified model to ",new_config_file)
+            self.tree.write(self.config_file)
+            # self.tree.write(new_config_file)  # saves modified XML to <output_dir>/config.xml 
+            # sys.exit(1)
 
             # Operate from tmpdir. XML: <folder>,</folder>; temporary output goes here.  May be copied to cache later.
-            tdir = os.path.abspath('tmpdir')
-            os.chdir(tdir)   # run exec from here on nanoHUB
+            if self.nanohub_flag:
+                tdir = os.path.abspath('tmpdir')
+                os.chdir(tdir)   # run exec from here on nanoHUB
             # sub.update(tdir)
             # subprocess.Popen(["../bin/myproj", "config.xml"])
 
@@ -183,6 +235,8 @@ class RunModel(QWidget):
         if self.vis_tab:
             # self.vis_tab.reset_axes()
             self.vis_tab.reset_model_flag = True
+            self.vis_tab.reset_plot_range()
+            self.vis_tab.init_plot_range(self.config_tab) # heaven help the person who needs to understand this
 
         # for f in Path('./output').glob('*.*'):
         #     try:
@@ -204,12 +258,9 @@ class RunModel(QWidget):
         if self.p is None:  # No process running.
             # self.vis_tab.setEnabled(True)
             # self.pStudio.enablePlotTab(True)
-            # self.tab_widget.enablePlotTab(True)
-
-            # self.tab_widget.setTabEnabled(5, True)
-            self.tab_widget.setTabEnabled(6, True)   # enable Plot tab 
-            self.tab_widget.setTabEnabled(7, True)   # enable Legend tab 
-
+            # self.pStudio.enableLegendTab(True)
+            self.tab_widget.setTabEnabled(6, True)   # enable (allow to be selected) the Plot tab
+            self.tab_widget.setTabEnabled(7, True)   # enable Legend tab
             self.message("Executing process")
             self.p = QProcess()  # Keep a reference to the QProcess (e.g. on self) while it's running.
             self.p.readyReadStandardOutput.connect(self.handle_stdout)
@@ -222,20 +273,24 @@ class RunModel(QWidget):
             if self.nanohub_flag:
                 self.p.start("submit",["--local",exec_str,xml_str])
             else:
-                print("run_tab.py: running: ",exec_str,xml_str)
+                logging.debug(f'\nrun_tab.py: running: {exec_str}, {xml_str}')
                 self.p.start(exec_str, [xml_str])
+
+                # print("\n\nrun_tab.py: running: ",exec_str," output/config.xml")
+                # self.p.start(exec_str, ["output/config.xml"])
             # self.p = None  # No, don't do this
 
             self.legend_tab.reload_legend()  # new, not sure about timing - creation vs. display
         else:
-            print("self.p is not None???")
+            logging.debug(f'self.p is not None???')
+            print(f'self.p is not None???')
 
     def cancel_model_cb(self):
-        print("===========  cancel_model_cb():  ============")
+        logging.debug(f'===========  cancel_model_cb():  ============')
         if self.p:  # process running.
-            # self.p.kill()
-            self.p.terminate()
-        self.run_button.setEnabled(True)
+            self.p.kill()
+            # self.p.terminate()
+            self.run_button.setEnabled(True)
 
     def handle_stderr(self):
         data = self.p.readAllStandardError()
@@ -255,10 +310,12 @@ class RunModel(QWidget):
         }
         state_name = states[state]
         self.message(f"State changed: {state_name}")
+        # self.message(f"Starting in a few secs...")   # only for "Starting"
 
     def process_finished(self):
         self.message("Process finished.")
-        print("-- process finished.")
+        # print("-- process finished.")
+        self.vis_tab.first_plot_cb("foo")
         if self.nanohub_flag:
             self.download_menu.setEnabled(True)
         self.p = None
